refactor: extract fetchWithTimeout to shared helper
Address Greptile review feedback: centralize timeout logic in
fetch-utils.ts instead of duplicating AbortController pattern.

- Add src/memory/fetch-utils.ts with shared fetchWithTimeout()
- Update batch-openai.ts to import from fetch-utils
- Update embeddings-openai.ts to use shared helper

This ensures consistent timeout behavior across both code paths.
