+++ src/memory/embeddings-openai.ts
@@ -1,7 +1,6 @@
 import type { EmbeddingProvider, EmbeddingProviderOptions } from "./embeddings.js";
 import { requireApiKey, resolveApiKeyForProvider } from "../agents/model-auth.js";
-
-const SYNC_FETCH_TIMEOUT_MS = 30_000;
+import { fetchWithTimeout, DEFAULT_FETCH_TIMEOUT_MS } from "./fetch-utils.js";
 
 export type OpenAiEmbeddingClient = {
   baseUrl: string;
@@ -33,19 +32,15 @@ export async function createOpenAiEmbeddingProvider(
     if (input.length === 0) {
       return [];
     }
-    const controller = new AbortController();
-    const timeout = setTimeout(() => controller.abort(), SYNC_FETCH_TIMEOUT_MS);
-    let res: Response;
-    try {
-      res = await fetch(url, {
+    const res = await fetchWithTimeout(
+      url,
+      {
         method: "POST",
         headers: client.headers,
         body: JSON.stringify({ model: client.model, input }),
-        signal: controller.signal,
-      });
-    } finally {
-      clearTimeout(timeout);
-    }
+      },
+      DEFAULT_FETCH_TIMEOUT_MS,
+    );
     if (!res.ok) {
       const text = await res.text();
       throw new Error(`openai embeddings failed: ${res.status} ${text}`);
